requirements.txt
click==8.1.3
hydra-core==1.3.1
Markdown==3.4.1
Pillow==9.4.0
python-dotenv==0.21.0
pytorch_lightning==1.8.6
PyYAML==6.0
setuptools==44.0.0
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
torchviz==0.0.2
tqdm==4.64.1
cookiecutter
flake8
isort
dvc
dvc[gdrive]
dvc[gs]
wandb
coverage
pre-commit
pytest
pytest-cov
